remove react-dates
