+++ package.json
@@ -13,7 +13,6 @@
     "@types/node": "^14.14.41",
     "@types/react": "^17.0.3",
     "@types/react-calendar": "^3.1.4",
-    "@types/react-dates": "^21.8.2",
     "@types/react-dom": "^17.0.3",
     "@typescript-eslint/eslint-plugin": "^4.22.0",
     "@typescript-eslint/parser": "^4.22.0",
@@ -33,7 +32,6 @@
     "react": "^17.0.2",
     "react-bootstrap": "^1.5.2",
     "react-calendar": "^3.4.0",
-    "react-dates": "^21.8.0",
     "react-dom": "^17.0.2",
     "react-hook-form": "^7.3.6",
     "sass": "^1.32.12",
